+++ hata/ext/plugin_auto_reloader/compatibility.py
@@ -1,4 +1,4 @@
-__all__ = ()
+__all__ = ('AUTO_RELOAD_SUPPORTED',)
 
 from sys import platform
 
@@ -7,8 +7,9 @@ AUTO_RELOAD_SUPPORTED = False
 INotify = None
 INotifyFlags = None
 
+IS_LINUX = platform == 'linux'
 
-if platform == 'linux':
+if IS_LINUX:
     try:
         from inotify_simple import INotify, flags as INotifyFlags
     except ImportError: