+++ hata/ext/plugin_auto_reloader/utils/update_auto_reloader_.py
@@ -1,4 +1,4 @@
-__all__ = ('update_auto_reload',)
+__all__ = ('update_auto_reloader',)
 
 from scarletio import copy_docs
 
@@ -6,5 +6,5 @@ from ..plugin_auto_reloader import PLUGIN_AUTO_RELOADER_MANAGER, PluginAutoReloa
 
 
 @copy_docs(PluginAutoReloaderManager.update)
-def update_auto_reload():
+def update_auto_reloader():
     return PLUGIN_AUTO_RELOADER_MANAGER.update()